+++ neovim/.config/nvim/lua/plugins/conform.lua
@@ -1,6 +1,7 @@
 return {
     'stevearc/conform.nvim',
-    event = { 'BufReadPre', 'BufNewFile' },
+    --event = { 'BufReadPre', 'BufNewFile' },
+    event = 'BufWritePre',
     cmd = { 'ConformInfo' },
     opts = {
         formatters_by_ft = {
@@ -13,9 +14,16 @@ return {
             lua = { 'stylua' },
             markdown = { 'prettier' },
             ruby = { 'rubocop' },
+            sh = { 'shfmt' },
             typescript = { 'prettier' },
             yaml = { 'prettier' },
             vue = { 'prettier' },
+        },
+        formatters = {
+            shfmt = {
+                command = 'shfmt',
+                prepend_args = { '-p', '-i', '0', '-sr', '-bn' },
+            },
         },
         -- Set default options
         default_format_opts = {